OSIS-629 End year for partim form
