+++ base/forms/learning_unit/external_learning_unit.py
@@ -32,7 +32,8 @@ from django.utils.translation import ugettext_lazy as _
 
 from base.forms.learning_unit.learning_unit_create import LearningUnitModelForm, LearningUnitYearModelForm, \
     LearningContainerModelForm, EntitiesVersionChoiceField, LearningContainerYearModelForm
-from base.forms.learning_unit.learning_unit_create_2 import LearningUnitBaseForm, merge_data
+from base.forms.learning_unit.learning_unit_create_2 import LearningUnitBaseForm
+from base.forms.learning_unit.learning_unit_partim import merge_data
 from base.forms.utils.acronym_field import ExternalAcronymField
 from base.models import entity_version
 from base.models.campus import Campus
