+++ base/models/learning_unit.py
@@ -25,9 +25,10 @@
 ##############################################################################
 from django.core.exceptions import ValidationError
 from django.db import models, IntegrityError
+from django.db.models import Max
 from django.utils.translation import ugettext_lazy as _
 
-from base.models.academic_year import current_academic_year
+from base.models.academic_year import current_academic_year, AcademicYear
 from base.models.enums.learning_unit_periodicity import PERIODICITY_TYPES, ANNUAL
 from base.models.enums.learning_unit_year_subtypes import PARTIM, FULL
 from base.models.proposal_learning_unit import ProposalLearningUnit
@@ -62,8 +63,9 @@ class LearningUnit(SerializableModel):
     external_id = models.CharField(max_length=100, blank=True, null=True)
     learning_container = models.ForeignKey('LearningContainer', blank=True, null=True)
     changed = models.DateTimeField(null=True, auto_now=True)
-    start_year = models.IntegerField()
-    end_year = models.IntegerField(blank=True, null=True)
+    start_year = models.IntegerField(_('start_year'))
+    end_year = models.IntegerField(blank=True, null=True, verbose_name=_('end_year_title'))
+    # TODO is it useful?
     progress = None
     periodicity = models.CharField(max_length=20, choices=PERIODICITY_TYPES, default=ANNUAL,
                                    verbose_name=_('periodicity'))
@@ -147,6 +149,14 @@ class LearningUnit(SerializableModel):
                 raise ValidationError(
                     {'periodicity': _('The periodicity of the parent and the partims do not match')}
                 )
+
+    @property
+    def max_end_year(self):
+        """ Compute the maximal possible end_year value when the end_year is None """
+        if self.end_year:
+            return self.end_year
+
+        return AcademicYear.objects.filter(learningunityear__learning_unit=self).aggregate(Max('year'))['year__max']
 
 
 def find_by_id(learning_unit_id):
