+++ base/templates/learning_unit/blocks/create_form_learning_unit.html
@@ -49,13 +49,19 @@
             {% bootstrap_field learning_container_year_form.campus form_group_class="col-md-3" %}
         </div>
         <div class="row">
-            {% bootstrap_field learning_unit_year_form.academic_year form_group_class="col-md-2" %}
             {% bootstrap_field learning_unit_form.periodicity form_group_class="col-md-2" %}
-            {% bootstrap_field learning_container_year_form.language form_group_class="col-md-2" %}
-            {% bootstrap_field learning_unit_year_form.credits form_group_class="col-md-1" %}
+            {% bootstrap_field learning_container_year_form.language form_group_class="col-md-3" %}
+            {% bootstrap_field learning_unit_year_form.credits form_group_class="col-md-2" %}
             {% bootstrap_field learning_unit_year_form.session form_group_class="col-md-2" %}
             {% bootstrap_field learning_unit_year_form.quadrimester form_group_class="col-md-2" %}
             {% bootstrap_field learning_unit_year_form.status form_group_class="col-md-1" %}
+        </div>
+            <br>
+        <div class="row">
+            {% bootstrap_field learning_unit_year_form.academic_year form_group_class="col-md-3" %}
+            {% if learning_unit_form.end_year %}
+                {% bootstrap_field learning_unit_form.end_year form_group_class="col-md-3" %}
+            {% endif %}
         </div>
     </div>
 </div>
