+++ base/tests/forms/test_learning_unit_partim.py
@@ -34,7 +34,8 @@ from django.utils.translation import ugettext_lazy as _
 from base.forms.learning_unit.learning_unit_create import EntityContainerBaseForm
 from base.forms.learning_unit.learning_unit_create import LearningUnitYearModelForm, \
     LearningUnitModelForm, LearningContainerYearModelForm, LearningContainerModelForm
-from base.forms.learning_unit.learning_unit_create_2 import PartimForm, PARTIM_FORM_READ_ONLY_FIELD
+from base.forms.learning_unit.learning_unit_partim import PARTIM_FORM_READ_ONLY_FIELD, PartimForm, \
+    LearningUnitPartimModelForm
 from base.forms.utils import acronym_field
 from base.models.enums import learning_unit_year_subtypes
 from base.models.enums.learning_unit_periodicity import ANNUAL, BIENNIAL_EVEN
@@ -101,7 +102,7 @@ class TestPartimFormInit(LearningUnitPartimFormContextMixin):
                               instance=wrong_instance)
 
     def test_model_forms_case_creation(self):
-        form_classes_expected = [LearningUnitModelForm, LearningUnitYearModelForm, LearningContainerModelForm,
+        form_classes_expected = [LearningUnitPartimModelForm, LearningUnitYearModelForm, LearningContainerModelForm,
                                  LearningContainerYearModelForm, EntityContainerBaseForm]
         form = _instanciate_form(learning_unit_full=self.learning_unit_year_full.learning_unit,
                                  academic_year=self.current_academic_year)
@@ -111,7 +112,7 @@ class TestPartimFormInit(LearningUnitPartimFormContextMixin):
     def test_inherit_initial_values(self):
         """This test will check if field are pre-full in by value of full learning unit year"""
         expected_initials = {
-            LearningUnitModelForm: {
+            LearningUnitPartimModelForm: {
                 'periodicity': self.learning_unit_year_full.learning_unit.periodicity
             },
             LearningUnitYearModelForm: {
@@ -166,7 +167,7 @@ class TestPartimFormInit(LearningUnitPartimFormContextMixin):
            in form_cls_to_validate"""
         partim_form = _instanciate_form(learning_unit_full=self.learning_unit_year_full.learning_unit,
                                         academic_year=self.learning_unit_year_full.academic_year)
-        expected_form_cls = [LearningUnitModelForm, LearningUnitYearModelForm]
+        expected_form_cls = [LearningUnitPartimModelForm, LearningUnitYearModelForm]
         self.assertEqual(partim_form.form_cls_to_validate, expected_form_cls)
 
 
@@ -273,7 +274,7 @@ class TestPartimFormIsValid(LearningUnitPartimFormContextMixin):
         self.assertTrue(form.is_valid(), form.errors)
 
     def _test_learning_unit_model_form_instance(self, partim_form, post_data):
-        form_instance = partim_form.forms[LearningUnitModelForm]
+        form_instance = partim_form.forms[LearningUnitPartimModelForm]
         fields_to_validate = ['faculty_remark', 'other_remark']
         self._assert_equal_values(form_instance.instance, post_data, fields_to_validate)
         # Periodicity inherit from parent [Cannot be modify by user]
@@ -338,7 +339,7 @@ class TestPartimFormSave(LearningUnitPartimFormContextMixin):
     """Unit tests for save() for save"""
     @mock.patch('base.forms.learning_unit.learning_unit_create.LearningUnitModelForm.save')
     @mock.patch('base.forms.learning_unit.learning_unit_create.LearningUnitYearModelForm.save')
-    @mock.patch('base.forms.learning_unit.learning_unit_create_2.PartimForm._get_entity_container_year',
+    @mock.patch('base.forms.learning_unit.learning_unit_partim.PartimForm._get_entity_container_year',
                 side_effect=lambda *args: [])
     def test_save_method_mocked(self, mock_get_entity_container_year, mock_luy_form_save, mock_lu_form_save):
         learning_container_year_full = self.learning_unit_year_full.learning_container_year
